+++ Craftsman/Commands/ListCommand.cs
@@ -29,6 +29,9 @@
             WriteHelpHeader(@$"{Environment.NewLine}    new");
             WriteHelpText(@$"      new:domain          Create a new DDD based project");
 
+            WriteHelpHeader(@$"{Environment.NewLine}    register");
+            WriteHelpText(@$"      register:consumer   Creates a new receive endpoint for event bus consumption.");
+
             WriteHelpHeader(@$"{Environment.NewLine}Options:");
             WriteHelpText(@$"   -h, --help          Display this help message");
 
@@ -39,6 +42,7 @@
             WriteHelpText(@$"   craftsman add:entity -h");
             WriteHelpText(@$"   craftsman add:message -h");
             WriteHelpText(@$"   craftsman add:property -h");
+            WriteHelpText(@$"   craftsman register:consumer -h");
         }
     }
 }