+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `AddBus` and `Broker` props to BC template
 - Added `add:message` command
 - Updated entity name and entity prop names first letter to always be capitalized
+- Added `register:consumer` command
 
 ## [0.9.3] - 2021-04-10